Handle deleted objects in sqlite queue
