+++ src/database/mysql/mysql_database.h
@@ -65,6 +65,8 @@ private:
     std::string quote(const std::string& value) const override;
 
     std::shared_ptr<SQLResult> select(const std::string& query) override;
+    void del(std::string_view tableName, const std::string& clause, const std::vector<int>& ids) override;
+    void exec(std::string_view tableName, const std::string& query, int objId) override;
     int exec(const std::string& query, bool getLastInsertId = false) override;
 
     void storeInternalSetting(const std::string& key, const std::string& value) override;
